@@ -5,14 +5,13 @@ namespace SoftCircuits.SimpleLogFile
     /// <summary>
     /// Argument for <see cref="LogFile.LineLogged"/> event.
     /// </summary>
-    public class LineLoggedEventArgs
-        : EventArgs
+    /// <param name="text">Specifies the text that was logged.</param>
+    public class LineLoggedEventArgs(string text)
+                : EventArgs
     {
-        public string Text { get; set; }
-
-        public LineLoggedEventArgs(string text)
-        {
-            Text = text;
-        }
+        /// <summary>
+        /// The text that was logged.
+        /// </summary>
+        public string Text { get; set; } = text;
     }
 }
